LIBITD-1359. Created "render_error_page" method
Created "render_error_page" method in application_controller to
simplify notifying the user of an error.

Also inserted a check in the orcid_controller to verify that a record
has not already been added for a user. This is needed in case the user
hits the "Back" button on their browser to return to the
"Terms of Service" page, and then hits "Agree" again.

Deleted cas_attribute_error and auth_code_callback_error pages, as they
are no longer needed, having been replaced by the "shared/error"
template.

https://issues.umd.edu/browse/LIBITD-1359