LIBITD-1364. Fixed ENV variables handling so Docker build succeeds
Added throw-away SECRET_KEY_BASE PROD_DATABASE_ADAPTER variables
so "assets:precompile" step in Dockerfile can succeed.

Adjusted CAS_SERVER_URL check in config/application.rb so that it is
not performed when running the "assets:precompile" step.

Added config/secrets.yml, because we want to use that in favor of
new Rails 5.2 credentials mechanism.

These changes are needed because we want to be able to build a Docker
container from a "clean" git checkout, which will not have a ".env"
file.

Removed credentials.yml.enc because it is not needed.

https://issues.umd.edu/browse/LIBITD-1364